requirements.txt
flask
pandas
numpy
tensorflow-cpu
scikit-learn
gensim
openpyxl
nltk
matplotlib
seaborn
flask-cors
pymongo
PyJWT
werkzeug
